[camera] move method definitions to cpp files
