+++ src/aliceVision/camera/DistortionFisheye1.hpp
@@ -28,22 +28,10 @@ public:
     DistortionFisheye1* clone() const override { return new DistortionFisheye1(*this); }
 
     /// Add distortion to the point p (assume p is in the camera frame [normalized coordinates])
-    Vec2 addDistortion(const Vec2 & p) const override
-    {
-        const double k1 = _distortionParams.at(0);
-        const double r = std::hypot(p(0), p(1));
-        const double coef = (std::atan(2.0 * r * std::tan(0.5 * k1)) / k1) / r;
-        return  p * coef;
-    }
+    Vec2 addDistortion(const Vec2 & p) const override;
 
     /// Remove distortion (return p' such that disto(p') = p)
-    Vec2 removeDistortion(const Vec2& p) const override
-    {
-        const double k1 = _distortionParams.at(0);
-        const double r = std::hypot(p(0), p(1));
-        const double coef = 0.5 * std::tan(r * k1) / (std::tan(0.5 * k1) * r);
-        return  p * coef;
-    }
+    Vec2 removeDistortion(const Vec2& p) const override;
 
     ~DistortionFisheye1() override  = default;
 };
