+++ src/aliceVision/camera/IntrinsicsScaleOffset.hpp
@@ -9,8 +9,6 @@
 
 #include "IntrinsicBase.hpp"
 
-#include <aliceVision/version.hpp>
-
 namespace aliceVision {
 namespace camera {
 
@@ -25,7 +23,7 @@ public:
     IntrinsicsScaleOffset(unsigned int w, unsigned int h,
                           double scaleX, double scaleY,
                           double offsetX, double offsetY) :
-        IntrinsicBase(w, h), 
+        IntrinsicBase(w, h),
         _scale(scaleX, scaleY), _offset(offsetX, offsetY)
     {
     }
@@ -34,15 +32,7 @@ public:
 
     void copyFrom(const IntrinsicsScaleOffset& other) { *this = other; }
 
-    bool operator==(const IntrinsicBase& otherBase) const override
-    {
-        if (!IntrinsicBase::operator==(otherBase))
-            return false;
-        if (typeid(*this) != typeid(otherBase))
-            return false;
-        const IntrinsicsScaleOffset& other = static_cast<const IntrinsicsScaleOffset&>(otherBase);
-        return _scale.isApprox(other._scale) && _offset.isApprox(other._offset);
-    }
+    bool operator==(const IntrinsicBase& otherBase) const override;
 
     void setScale(const Vec2& scale) { _scale = scale; }
 
@@ -55,154 +45,46 @@ public:
     /**
      * @brief Principal point in image coordinate ((0,0) is image top-left).
      */
-    inline const Vec2 getPrincipalPoint() const 
+    inline const Vec2 getPrincipalPoint() const
     {
         Vec2 ret = _offset;
-        ret(0) += double(_w) * 0.5;
-        ret(1) += double(_h) * 0.5;
+        ret(0) += static_cast<double>(_w) * 0.5;
+        ret(1) += static_cast<double>(_h) * 0.5;
 
         return ret;
     }
 
     // Transform a point from the camera plane to the image plane
-    Vec2 cam2ima(const Vec2& p) const override
-    {
-        return p.cwiseProduct(_scale) + getPrincipalPoint();
-    }
+    Vec2 cam2ima(const Vec2& p) const override;
 
-    virtual Eigen::Matrix2d getDerivativeCam2ImaWrtScale(const Vec2& p) const
-    {
-        Eigen::Matrix2d M = Eigen::Matrix2d::Zero();
+    virtual Eigen::Matrix2d getDerivativeCam2ImaWrtScale(const Vec2& p) const;
 
-        M(0, 0) = p(0);
-        M(1, 1) = p(1);
-            
-        return M;
-    }
+    virtual Eigen::Matrix2d getDerivativeCam2ImaWrtPoint() const;
 
-    virtual Eigen::Matrix2d getDerivativeCam2ImaWrtPoint() const
-    {
-        Eigen::Matrix2d M = Eigen::Matrix2d::Zero();
-
-        M(0, 0) = _scale(0);
-        M(1, 1) = _scale(1);
-
-        return M;
-    }
-
-    virtual Eigen::Matrix2d getDerivativeCam2ImaWrtPrincipalPoint() const
-    {
-        return Eigen::Matrix2d::Identity();
-    }
+    virtual Eigen::Matrix2d getDerivativeCam2ImaWrtPrincipalPoint() const;
 
     // Transform a point from the image plane to the camera plane
-    Vec2 ima2cam(const Vec2& p) const override
-    {
-        Vec2 np;
+    Vec2 ima2cam(const Vec2& p) const override;
 
-        Vec2 pp = getPrincipalPoint();
+    virtual Eigen::Matrix<double, 2, 2> getDerivativeIma2CamWrtScale(const Vec2& p) const;
 
-        np(0) = (p(0) - pp(0)) / _scale(0);
-        np(1) = (p(1) - pp(1)) / _scale(1);
+    virtual Eigen::Matrix2d getDerivativeIma2CamWrtPoint() const;
 
-        return np;
-    }
-
-    virtual Eigen::Matrix<double, 2, 2> getDerivativeIma2CamWrtScale(const Vec2& p) const
-    {
-        Eigen::Matrix2d M = Eigen::Matrix2d::Zero();
-
-        Vec2 pp = getPrincipalPoint();
-
-        M(0, 0) = -(p(0) - pp(0)) / (_scale(0) * _scale(0));
-        M(1, 1) = -(p(1) - pp(1)) / (_scale(1) * _scale(1));
-
-        return M;
-    }
-
-    virtual Eigen::Matrix2d getDerivativeIma2CamWrtPoint() const
-    {
-        Eigen::Matrix2d M = Eigen::Matrix2d::Zero();
-
-        M(0, 0) = 1.0 / _scale(0);
-        M(1, 1) = 1.0 / _scale(1);
-
-        return M;
-    }
-
-    virtual Eigen::Matrix2d getDerivativeIma2CamWrtPrincipalPoint() const
-    {
-        Eigen::Matrix2d M = Eigen::Matrix2d::Zero();
-
-        M(0, 0) = - 1.0 / _scale(0);
-        M(1, 1) = - 1.0 / _scale(1);
-
-        return M;
-    }
+    virtual Eigen::Matrix2d getDerivativeIma2CamWrtPrincipalPoint() const;
 
     /**
      * @brief Rescale intrinsics to reflect a rescale of the camera image
      * @param factor a scale factor
      */
-    void rescale(float factor) override
-    {
-        IntrinsicBase::rescale(factor);
-
-        _scale *= factor;
-        _offset *= factor;
-    }
+    void rescale(float factor) override;
 
     // Data wrapper for non linear optimization (update from data)
-    bool updateFromParams(const std::vector<double>& params) override
-    {
-        if (params.size() != 4)
-        {
-            return false;
-        }
-
-        _scale(0) = params[0];
-        _scale(1) = params[1];
-        _offset(0) = params[2];
-        _offset(1) = params[3];
-
-        return true;
-    }
+    bool updateFromParams(const std::vector<double>& params) override;
 
     /**
      * @brief Import a vector of params loaded from a file. It is similar to updateFromParams but it deals with file compatibility.
      */
-    bool importFromParams(const std::vector<double>& params, const Version & inputVersion) override
-    {
-        std::vector<double> paramsLocal;
-        if (inputVersion < Version(1, 2, 0))
-        {
-            paramsLocal.resize(params.size() + 1);
-            paramsLocal[0] = params[0];
-            paramsLocal[1] = params[0];
-
-            for (int i = 1; i < params.size(); i++)
-            {
-                paramsLocal[i + 1] = params[i];
-            }
-        }
-        else
-        {
-            paramsLocal = params;
-        }
-
-        if (!updateFromParams(paramsLocal))
-        {
-            return false;
-        }
-
-        if (inputVersion < Version(1, 2, 1))
-        {
-            _offset(0) -= double(_w) / 2.0;
-            _offset(1) -= double(_h) / 2.0;
-        }
-
-        return true;
-    }
+    bool importFromParams(const std::vector<double>& params, const Version & inputVersion) override;
 
     /**
      * @brief Set initial Scale (for constraining minimization)
@@ -219,9 +101,9 @@ public:
      * @brief lock the ratio between fx and fy
      * @param lock is the ratio locked
      */
-    void setRatioLocked(bool lock) { _ratioLocked = lock; }
+    inline void setRatioLocked(bool lock) { _ratioLocked = lock; }
 
-    bool isRatioLocked() const { return _ratioLocked; }
+    inline bool isRatioLocked() const { return _ratioLocked; }
 
 protected:
     Vec2 _scale{1.0, 1.0};
